+++ mkenum/example_test.go
@@ -1,0 +1,14 @@
+package main
+
+//go:generate mkenum -t Weekday .
+type Weekday string
+
+const (
+	Monday    Weekday = "Monday"
+	Tuesday   Weekday = "Tuesday"
+	Wednesday Weekday = "Wednesday"
+	Thursday  Weekday = "Thursday"
+	Friday    Weekday = "Friday"
+	Saturday  Weekday = "Saturday"
+	Sunday    Weekday = "Sunday"
+)