+++ mkenum/mkenum_test.go
@@ -11,16 +11,3 @@ func Test_mkenum(t *testing.T) {
 	out, _ := exec.Command("go", "generate", ".").Output()
 	golden.Assert(t, string(out))
 }
-
-//go:generate mkenum -t Weekday .
-type Weekday string
-
-const (
-	Monday    Weekday = "Monday"
-	Tuesday   Weekday = "Tuesday"
-	Wednesday Weekday = "Wednesday"
-	Thursday  Weekday = "Thursday"
-	Friday    Weekday = "Friday"
-	Saturday  Weekday = "Saturday"
-	Sunday    Weekday = "Sunday"
-)